==> vaxMada/code/0_helper_functions.R <==
translate_question <- function(ch) {
  return(switch(EXPR=ch, 
                "Fanadiadiagna VAKSINY"="Vaccination survey", "FIRY AKOHO JIABY NAHAZO VAKSINY"="Total number of chicken vaccinated", "FIRY AKOHO JIABY TOMPINY"="total number of chicken owned",
                "Firy isan'ny ZANAKA AKOHO NAHAZO VAKSINY"="total number of chicks vaccinated", "Firy isan'ny AKOHO VANTONY NAHAZO VAKSINY"="total number of juvenile vaccinated", "Firy ny isan'ny AKOHO VAVY BE NAHAZO VAKSINY"="total number of hen vaccinated",
                "Firy ny isan'ny AKOHO LAHY BE NAHAZO VAKSINY"="total number of rooster vaccinated", "FIRY AKOHO TSY NAHAZO VAKSINY (TSY MILA)"="number of chickens not vaccinated (Do not want vaccination)", "FIRY AKOHO TSY NAHAZO VAKSINY (TSY NAHAZO)"="number of chickens not vaccinated (want vaccination)",
                "FIRY AKOHO NAHAZO VAKSINY"="number of chickens vaccinated", "VOLA NAHAZO"="Amount earned", "NOTE VAKSINY"="Comments about vaccine",
                "Fanadiadiagna AKOHO"="Chicken Survey", "Nandray anjara tamin'ny vaksiny farany? (Eny na Tsia)"="Participated to the previous vaccination (Yes/no)", "Firy isan'ny akoho JIABY amin'ny andro ake andre trano iraika?"="Current total number of  chickens at the household",
                "Firy isan'ny ZANAKA AKOHO jiaby amin'ny andro ake?"="Current total number of chicks", "Firy isan'ny AKOHO VANTONY amin'ny andro ake?"="Current total number of juveniles", "Firy ny isan'ny AKOHO VAVY BE amin'ny andro ake?"="Current total number of hens",
                "Firy ny isan'ny AKOHO LAHY BE amin'ny andro ake?"="Current total number of roosters", "Firy isan'ny akoho jiaby nohanina anatiny 4 vol lasa/tafaran'ny fanadihadiana/vaksiny farany?"="Total number of chickens eaten after the last vaccination survey (every 4 months)", "Firy isan'ny ZANAKA AKOHO nohanina tafaran'ny 4 volana/fanadihadiana/vaksiny farany?"="Total number of chicks eaten after the last vaccination survey",
                "Firy isan'ny AKOHO VANTONY nohanina tafaran'ny 4 volana/fanadihadiana/vaksiny farany?"="Total number of juveniles eaten after the last vaccination survey", "Firy isan'ny AKOHO VAVY BE nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of hens eaten after the last vaccination survey", "Firy isan'ny AKOHO LAHY BE nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of rooster eaten after the last vaccination survey",
                "Nisy akoho nohaninatafaran'ny fanadihadiana/vaksiny farany ka natao VAKSINY tanatin'ny 6 VOLANA LASA?"="Any chickens (vaccinated during the last 6 months) eaten after the last vaccination survey", "Raha eny, firy?"="If yes, how many? (Undefined)", "Firy isan'ny akoho jiaby namidy tafaran'ny fanadihadiana farany natao?"="Total number of chickens sold after the last survey",
                "Firy isan'ny ZANAKO AKOHO namidy tafaran'ny fanadihadiana farany natao?"="Total number of chicks sold after the last survey", "Otrinona no namarotana ZANAKA AKOHO IRAY? (ARIARY)"="Price of one chicks", "Firy isan'ny AKOHO VANTONY namidy tafaran'ny fanadihadiana farany natao?"="Total number of juveniles sold after the last survey",
                "Otrinona no namarotana AKOHO VANTONY IRAY? (ARIARY)"="Price of one juvenile", "Firy isan'ny AKOHO VAVY BE namidy tafaran'ny fanadihadiana farany natao?"="Total number of hens sold after the last survey", "Otrinona no namarotana AKOHO VAVY BE IRAY? (ARIARY)"="Price of one hen",
                "Firy isan'ny AKOHO LAHY BE namidy tafaran'ny fanadihadiana farany natao?"="Total number of roosters sold after the last survey", "Otrinona no namarotana AKOHO LAHY BE IRAY? (ARIARY)"="price of one rooster", "Nisy akoho namidy tamin'ny VOLANA LASA ka natao VAKSINY tanatin'ny 6 VOLANA LASA?"="Any chickens vaccinated in the last 6 months sold after the last vaccination survey",
                "Firy ny isan'ny akoho jiaby novidiana tafaran'ny fanadihadiana farany natao?"="Total number of chicken purchased after the last survey", "Otrinona vidin'ny ZANAKA AKOHO IRAY? (ARIARY)"="Price of one chick",
                "Otrinona vidin'ny AKOHO VANTONY IRAY? (ARIARY)"="Price of one juvenile", "Otrinona vidin'ny AKOHO VAVY BE IRAY? (ARIARY)"="price of one hen", "Otrinona vidin'ny AKOHO LAHY BE IRAY? (ARIARY)"="price of one rooster",
                "Firy isan'ny akoho jiaby natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Total number of chicken died by disease after the last survey", "Firy ZANAKA AKOHO natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Number of Chicks died after the last survey", "Firy AKOHO VANTONY natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Number of juveniles died after the last survey",
                "Firy AKOHO VAVY BE natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Number of hens died after the last survey", "Firy AKOHO LAHY BE natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Number of rooster died after the last survey", "Nisy akoho naty tafaran'ny fanadihadiana farany ka natao VAKSINY tanatin'ny 6 VOLANA LASA?"="Any chicken vaccinated in the last 6 months died after the last survey",
                "Firy isan'ny akoho jiaby azo KADO tafaran'ny fanadihadiana farany natao?"="Total number of  chickens got as gift after the last survey", "Firy isan'ny akoho jiaby NOMENA KADO tafaran'ny fanadihadiana farany natao?"="Total number of chickens gave as gifts after the last survey",
                "Firy isan'ny akoho jiany natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of chicken exchanged with another thing after the last survey", "Firy isan'ny ZANAKA AKOHO natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of chick exchanged with another thing after the last survey", "Firy isan'ny AKOHO VANTONY natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of juvenile exchanged with another thing after the last survey",
                "Firy isan'ny AKOHO VAVY BE natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of hen exchanged with another thing after the last survey", "Firy isan'ny AKOHO LAHY BE natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of rooster exchanged with another thing after the last survey", "Firy isan'ny akoho jiaby natao tsaboraha na nohanina tamin'ny asa tamin'olo?"="Total number of chickens eaten for exhumation or for community work",
                "Firy isan'ny ZANAKA AKOHO natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of chick eaten for exhumation or for community work", "Firy isan'ny AKOHO VANTONY natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of juvenile eaten for exhumation or for community work", "Firy isan'ny AKOHO VAVY BE natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of hen eaten for exhumation or for community work",
                "Firy isan'ny AKOHO LAHY BE natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of rooster eaten for exhumation or for community work", "Firy isan'ny akoho jiaby nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of chicken stolen after the last survey", "Firy isan'ny ZANAKA AKOHO nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of chick stolen after the last survey",
                "Firy isan'ny AKOHO VANTONY nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of juvenile stolen after the last survey", "Firy isan'ny AKOHO VAVY BE nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of hen stolen after the last survey", "Firy isan'ny AKOHO LAHY BE nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of rooster stolen after the last survey",
                "Firy isan'ny akoho jiaby lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of chicken eaten by snake or other predators after the last survey", "Firy isan'ny ZANAKA AKOHO lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of chick eaten by snake or other predators after the last survey", "Firy isan'ny AKOHO VANTONY lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of juvenile eaten by snake or other predators after the last survey",
                "Firy isan'ny AKOHO VAVY BE lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of hen eaten by snake or other predators after the last survey", "Firy isan'ny AKOHO LAHY BE lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of rooster eaten by snake or other predators after the last survey", "Firy isan'ny AKOHO VAVY BE nikotrika atody tafaran'ny fanadihadiana farany?"="Number of hen brooding eggs after the last survey",
                "Firy isan'ny ZANAKA AKOHO niboaka tafaran'ny fanadihadiana farany?"="Number of chick hatched after the last survey", "Firy isan'ny ZANAKA AKOHO niboaka ka naty tafaran'ny fanadihadiana farany?"="Number of chick hatched but died after the last survey", "Firy isan'ny atodin' AKOHO nohanina tafaran'ny fanadihadiana farany?"="Number of egg eaten after the last survey",
                "Firy isan'ny atodin' AKOHO namidy tafaran'ny fanadihadiana/vaksiny farany?"="Number of egg sold after the survey", "Otrinona no nandafosana ny atodin' AKOHO IRAY? (ARIARY)"="Price of one egg if sold", "Firy mety isan'ny vaksiny ilaina amin'ny manaraka?"="Total of the vaccine doses needed for the next campaign",
                "Fanadiadiagna VORONA"="Palmiped Survey", "Firy vorondrano tompony andre trano iraika?"="Number of palmiped at the household", "Firy vorondrano NIHININA andre trano iraika tafaran'ny fanadihadiana/vaksiny farany?"="Number of palmiped eaten at the household after the last survey",
                "Firy vorondrano NANDAFO andre trano iraika tafaran'ny fanadihadiana/vaksiny farany?"="Number of palmiped sold at the household after the last survey", "Firy atody vorona NIHININA andre trano iraika tafaran'ny fanadihadiana/vaksiny farany?"="Number of palmiped egg eaten at the household after the last survey", "Firy atody vorona NANDAFO andre trano iraika tafaran'ny fanadihadiana/vaksiny farany?"="Number of palmiped egg sold at the household after the last survey",
                "Firy atody vorona MAMANA andre trano iraika tafaran'ny fanadihadiana/vaksiny farany?"="Number of palmiped brooding after the last survey", "Firy atody vorona FOY andre trano iraika tafaran'ny fanadihadiana/vaksiny farany?"="Number of palmiped egg hatched after the last survey", "Firy vorondrano NATY andre trano iraika tafaran'ny fanadihadiana/vaksiny farany?"="Number of palmiped died after the last survey",
                "Vorondrano: Naty avy inona?"="Cause of the death of the palmiped", "Vidiny akoho na vorondrano iraika ha nandafo atafaran'ny fanadihadiana/vaksiny farany? OHATRA ka tsy nandafo- aza manoratra raha"="Price  of one chicken or palmiped sold after the last survey or vaccination? If nothing sold, leave it blank", "Vidiny atody vorona iraika ha nandafo tafaran'ny fanadihadiana/vaksiny farany? OHATRA ka tsy nandafo- aza manoratra raha"="Price  of one palmiped egg sold after the last survey or vaccination? If nothing sold, leave it blank",
                "Note"="Note",
                "Raha eny, firy?...29"="If yes, how many? (4)", "Raha eny, firy?...40"="If yes, how many? (5)", "Raha eny, firy?...52"="If yes, how many? (6)",
                # Translated second time 9-13-22
                "Fanadiadiagna AKOHO SY VAKSINY"="Vaccination survey", "Firy akoho jiaby?"="total number of chicken owned",
                "ZANAKA AKOHO natao vaksiny ?"="total number of chicks vaccinated",  "AKOHO VANTONY natao vaksiny ?"="total number of juvenile vaccinated",
                "AKOHO VAVY BE natao vaksiny ?"="total number of hen vaccinated", "AKOHO LAHY BE natao vaksiny ?"="total number of rooster vaccinated",
                "Firy akoho tsy mila vaksiny?"="number of chickens not vaccinated (Do not want vaccination)", "Firy akoho tsy vita vaksiny tsy takatra na tsy hita?"="number of chickens not vaccinated (want vaccination)",
                "Firy akoho vita vaksiny?"="number of chickens vaccinated", "Ariary nagnefa"="Amount earned",
                "Jareo nifindra na tsisy olo an'trano? (Nifindra/Tsisy olo ako)"="Household moved or people not catched",
                "Fanamarihana"="Comments about vaccine", "Fanadiadiagna AKOHO SY VAKSINY"="Survey chicken and vaccination",
                "Nanagno vaksiny Akoho tamin'ny AUG 2020?"="Participated to the previous vaccination Aug 2020 (Yes/no)", "Firy isan'ny akoho JIABY-gnao amin'ny andro ake?"="Current total number of  chickens at the household",
                "Firy isan'ny ZANAKA AKOHO jiaby amin'ny andro ake?"="Current total number of chicks", "Firy isan'ny AKOHO VANTONY amin'ny andro ake?"="Current total number of juveniles",
                "Firy ny isan'ny AKOHO VAVY BE amin'ny andro ake?"="Current total number of hens", "Firy ny isan'ny AKOHO LAHY BE amin'ny andro ake?"="Current total number of roosters",
                "Firy isan'ny akoho jiaby nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of chickens eaten after the last vaccination survey", "Firy isan'ny akoho jiaby nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of chickens eaten after the last vaccination survey",
                "Firy isan'ny ZANAKA AKOHO nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of chicks eaten after the last vaccination survey", "Firy isan'ny ZANAKA AKOHO nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of chicks eaten after the last vaccination survey",
                "Firy isan'ny AKOHO VANTONY nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of juveniles eaten after the last vaccination survey", "Firy isan'ny AKOHO VAVY BE nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of hens eaten after the last vaccination survey",
                "Firy isan'ny AKOHO LAHY BE nohanina tafaran'ny fanadihadiana/vaksiny farany?"="Total number of rooster eaten  after the last vaccination survey", "Nisy akoho nohaninatafaran'ny fanadihadiana/vaksiny farany ka natao VAKSINY tanatin'ny 6 VOLANA LASA?"="Any chickens (vaccinated during the last 6 months) eaten after the last vaccination survey",
                "Raha eny, firy?...16"="If yes, how many? (7)", "Firy isan'ny akoho jiaby namidy tafaran'ny fanadihadiana farany natao?"="Total number of chickens sold after the last survey",
                "Firy isan'ny ZANAKO AKOHO namidy tafaran'ny fanadihadiana farany natao?"="Total number of chicks sold after the last survey", "Otrinona no namarotana ZANAKA AKOHO IRAY? (ARIARY)"="Price of one chicks",
                "Firy isan'ny AKOHO VANTONY namidy tafaran'ny fanadihadiana farany natao?"="Total number of juveniles sold after the last survey", "Otrinona no namarotana AKOHO VANTONY IRAY? (ARIARY)"="Price of one juvenile",
                "Firy isan'ny AKOHO VAVY BE namidy tafaran'ny fanadihadiana farany natao?"="Total number of hens sold after the last survey", "Otrinona no namarotana AKOHO VAVY BE IRAY? (ARIARY)"="Price of one hen",
                "Firy isan'ny AKOHO LAHY BE namidy tafaran'ny fanadihadiana farany natao?"="Total number of roosters sold after the last survey", "Otrinona no namarotana AKOHO LAHY BE IRAY? (ARIARY)"="price of one rooster",
                "Nisy akoho namidy tamin'ny VOLANA LASA ka natao VAKSINY tanatin'ny 6 VOLANA LASA?"="Any chickens vaccinated in the last 6 months sold after the last vaccination survey", "Raha eny, firy?...27"="If yes, how many? (8)",
                "Firy ny isan'ny akoho jiaby novidiana tafaran'ny fanadihadiana farany natao?"="Total number of chicken purchased after the last survey", "Avy taiza toerana nihavian'ny akoho novidignao?"="Place or origin of the puirchased chicken",
                "Otrinona vidin'ny ZANAKA AKOHO IRAY? (ARIARY)"="Price of one chick", "Otrinona vidin'ny AKOHO VANTONY IRAY? (ARIARY)"="Price of one juvenile",
                "Otrinona vidin'ny AKOHO VAVY BE IRAY? (ARIARY)"="price of one hen", "Otrinona vidin'ny AKOHO LAHY BE IRAY? (ARIARY)"="price of one rooster",
                "Firy isan'ny akoho jiaby natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Total number of chicken died by disease after the last survey", "Ino aretina nahafaty ny akognao?"="What disease killed  the chicken",
                "Firy ZANAKA AKOHO natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Number of Chicks died after the last survey", "Firy AKOHO VANTONY natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Number of juveniles died after the last survey",
                "Firy AKOHO VAVY BE natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Number of hens died after the last survey", "Firy AKOHO LAHY BE natin'ny aretina tafaran'ny fanadihadiana farany natao?"="Number of rooster died after the last survey",
                "Nisy akoho naty tafaran'ny fanadihadiana farany ka natao VAKSINY tanatin'ny 6 VOLANA LASA?"="Any chicken vaccinated in the last 6 months died after the last survey", "Raha eny, firy?...41"= "If yes, how many? (9)",
                "Firy isan'ny akoho jiaby azo KADO tafaran'ny fanadihadiana farany natao?"="Total number of  chickens got as gift after the last survey", "Avy taiza toerana nihavian'ny akoho azo kadô?"="place or origine of the chicken got as gift",
                "Firy isan'ny akoho jiaby NOMENA KADO tafaran'ny fanadihadiana farany natao?"="Total number of chickens gave as gifts after the last survey", "Firy isan'ny akoho jiany natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of chicken exchanged with another thing after the last survey",
                "Firy isan'ny ZANAKA AKOHO natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of chick exchanged with another thing after the last survey", "Firy isan'ny AKOHO VANTONY natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of juvenile exchanged with another thing after the last survey",
                "Firy isan'ny AKOHO VAVY BE natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of hen exchanged with another thing after the last survey", "Firy isan'ny AKOHO LAHY BE natakalo raha hafa tafaran'ny fanadihadiana farany natao?"="Total number of rooster exchanged with another thing after the last survey",
                "Firy isan'ny akoho jiaby natao tsaboraha na nohanina tamin'ny asa tamin'olo?"="Total number of chickens eaten for exhumation or for community work", "Firy isan'ny ZANAKA AKOHO natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of chick eaten for exhumation or for community work",
                "Firy isan'ny ZANAKA AKOHO natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of chick eaten for exhumation or for community work", "Firy isan'ny AKOHO VANTONY natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of juvenile eaten for exhumation or for community work",
                "Firy isan'ny AKOHO VANTONY natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of juvenile eaten for exhumation or for community work", "Firy isan'ny AKOHO VAVY BE natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of hen eaten for exhumation or for community work",
                "Firy isan'ny AKOHO VAVY BE natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of hen eaten for exhumation or for community work", "Firy isan'ny AKOHO LAHY BE natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of rooster eaten for exhumation or for community work",
                "Firy isan'ny AKOHO LAHY BE natao tsaboraha na nohanina  tamin'ny asa tamin'olo tafaran'ny fanadihadiana farany?"="Total number of rooster eaten for exhumation or for community work", "Firy isan'ny akoho jiaby nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of chicken stolen after the last survey",
                "Firy isan'ny ZANAKA AKOHO nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of chick stolen after the last survey", "Firy isan'ny AKOHO VANTONY nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of juvenile stolen after the last survey",
                "Firy isan'ny AKOHO VAVY BE nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of hen stolen after the last survey", "Firy isan'ny AKOHO LAHY BE nisy nangalatra tafaran'ny fanadihadiana farany?"="Total number of rooster stolen after the last survey",
                "Firy isan'ny akoho jiaby lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of chicken eaten by snake or other predators after the last survey", "Firy isan'ny ZANAKA AKOHO lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of chick eaten by snake or other predators after the last survey",
                "Firy isan'ny AKOHO VANTONY lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of juvenile eaten by snake or other predators after the last survey", "Firy isan'ny AKOHO VAVY BE lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of hen eaten by snake or other predators after the last survey",
                "Firy isan'ny AKOHO LAHY BE lanin'ny bibilava na bibidia tafaran'ny fanadihadiana farany?"="Total number of roaster eaten by snake or other predators after the last survey", "Firy isan'ny AKOHO VAVY BE nikotrika atody tafaran'ny fanadihadiana farany?"="Number of hen brooding eggs after the last survey",
                "Mety firy ake ake ny isan'ny atody akotriky ny akoho vavy iraiky?"="Number of eggs per hen brooding", "Firy isan'ny ZANAKA AKOHO niboaka tafaran'ny fanadihadiana farany?"="Number of chick hatched after the last survey",
                "Firy isan'ny ZANAKA AKOHO niboaka ka naty tafaran'ny fanadihadiana farany?"="Number of chick hatched but died after the last survey", "Firy isan'ny atodin' AKOHO nohanina tafaran'ny fanadihadiana farany?"="Number of egg eaten after the last survey",
                "Firy isan'ny atodin' AKOHO namidy tafaran'ny fanadihadiana/vaksiny farany?"="Number of egg sold after the survey", "Otrinona no nandafosana ny atodin' AKOHO IRAY? (ARIARY)"="Price of one egg if sold",
                "Firy mety isan'ny vaksiny ilaina amin'ny manaraka?"="Total of the vaccine doses needed for the next campaign",
                "Nanagno vaksiny Akoho tamin'ny AVRIL 2021?"="Participated to the previous vaccination April 2021 (Yes/no)",
                "Tanana"='village', "Tokantrano"='Household',
                ch))}
switch_date <- function(ch) {
  return(switch(EXPR=ch, 
                "2017_Janvier"=as.Date('01-01-2017', format="%m-%d-%Y"), "2016_Septembre"=as.Date('09-01-2016', format="%m-%d-%Y"), # Makira ownership
                "2016_Octobre"=as.Date('10-01-2016', format="%m-%d-%Y"),"2016_Novembre"=as.Date('11-01-2016', format="%m-%d-%Y"),
                "2016_Decembre"=as.Date('12-01-2016', format="%m-%d-%Y"), "2016_Janvier"=as.Date('01-01-2016', format="%m-%d-%Y"),
                "2016_Fevrier"=as.Date('02-01-2016', format="%m-%d-%Y"), "2016_Mars"=as.Date('03-01-2016', format="%m-%d-%Y"),
                "2016_Avril"=as.Date('04-01-2016', format="%m-%d-%Y"), "2016_Mai"=as.Date('05-01-2016', format="%m-%d-%Y"),
                "2016_Juin"=as.Date('06-01-2016', format="%m-%d-%Y"), "2016_Juillet"=as.Date('07-01-2016', format="%m-%d-%Y"),
                "2016_Aout"=as.Date('08-01-2016', format="%m-%d-%Y"), "2015_JANVIER"=as.Date('01-01-2015', format="%m-%d-%Y"),
                "2015_FEVRIER"=as.Date('02-01-2015', format="%m-%d-%Y"), "2015_MARS"=as.Date('03-01-2015', format="%m-%d-%Y"),
                "2015_AVRIL"=as.Date('04-01-2015', format="%m-%d-%Y"), "2015_MAI"=as.Date('05-01-2015', format="%m-%d-%Y"),
                "2015_JUIN"=as.Date('06-01-2015', format="%m-%d-%Y"), "2015_JUILLET"=as.Date('07-01-2015', format="%m-%d-%Y"),
                "2015_Aout"=as.Date('08-01-2015', format="%m-%d-%Y"), "2015_Septembre"=as.Date('09-01-2015', format="%m-%d-%Y"),
                "2015_Octobre"=as.Date('10-01-2015', format="%m-%d-%Y"), "2015_Novembre"=as.Date('11-01-2015', format="%m-%d-%Y"),
                "2015_Decembre"=as.Date('12-01-2015', format="%m-%d-%Y"), "2015_septembre"=as.Date('09-01-2015', format="%m-%d-%Y"),
                "2015_octobre"=as.Date('10-01-2015', format="%m-%d-%Y"), "2014_Fevrier"=as.Date('02-01-2014', format="%m-%d-%Y"),
                "2014_mars"=as.Date('03-01-2014', format="%m-%d-%Y"), "2014_avril"=as.Date('04-01-2014', format="%m-%d-%Y"),
                "2014_mai "=as.Date('05-01-2014', format="%m-%d-%Y"), "2014_juin"=as.Date('06-01-2014', format="%m-%d-%Y"),
                "2014_juillet"=as.Date('07-01-2014', format="%m-%d-%Y"), "2014_aout"=as.Date('08-01-2014', format="%m-%d-%Y"),
                "2014_septembre"=as.Date('09-01-2014', format="%m-%d-%Y"), "2014_octobre"=as.Date('10-01-2014', format="%m-%d-%Y"),
                "2014_Novembre"=as.Date('11-01-2014', format="%m-%d-%Y"), "2014_decembre"=as.Date('12-01-2014', format="%m-%d-%Y"),
                "2013_Janvier"=as.Date('01-01-2013', format="%m-%d-%Y"), "2013_Fevrier"=as.Date('02-01-2013', format="%m-%d-%Y"),
                "2013_Avril"=as.Date('04-01-2013', format="%m-%d-%Y"), "2013_Mai"=as.Date('05-01-2013', format="%m-%d-%Y"), 
                "2013_Mars"=as.Date('03-01-2013', format="%m-%d-%Y"), "2013_Juin"=as.Date('06-01-2013', format="%m-%d-%Y"),
                "2013_juillet"=as.Date('07-01-2013', format="%m-%d-%Y"), "2013_Aout"=as.Date('08-01-2013', format="%m-%d-%Y"),
                "2013_Decembre"=as.Date('12-01-2013', format="%m-%d-%Y"), "2013_octobre"=as.Date('10-01-2013', format="%m-%d-%Y"),
                "2013_septembre"=as.Date('09-01-2013', format="%m-%d-%Y"), "2013_Novembre"=as.Date('11-01-2013', format="%m-%d-%Y"),
                "2012_Janvier"=as.Date('01-01-2012', format="%m-%d-%Y"), "2012_Fevrier"=as.Date('02-01-2012', format="%m-%d-%Y"),
                "2012_Mars"=as.Date('03-01-2012', format="%m-%d-%Y"), "2012_AVRIL"=as.Date('04-01-2012', format="%m-%d-%Y"),
                "2012_MAI"=as.Date('05-01-2012', format="%m-%d-%Y"), "2012_juin"=as.Date('06-01-2012', format="%m-%d-%Y"),  
                "2012_juillet"=as.Date('07-01-2012', format="%m-%d-%Y"), "2012_Aout"=as.Date('08-01-2012', format="%m-%d-%Y"),
                "2012_Septembre"=as.Date('09-01-2012', format="%m-%d-%Y"), "2012_octobre"=as.Date('10-01-2012', format="%m-%d-%Y"),
                "2012_Novembre"=as.Date('11-01-2012', format="%m-%d-%Y"), "2012_Decembre"=as.Date('12-01-2012', format="%m-%d-%Y"),
                "2011_AOUT"=as.Date('08-01-2011', format="%m-%d-%Y"), "2011_SEPTEMBRE"=as.Date('09-01-2011', format="%m-%d-%Y"),
                "2011_OCTOBRE"=as.Date('10-01-2011', format="%m-%d-%Y"), "2011_Novembre"=as.Date('11-01-2011', format="%m-%d-%Y"),
                "2011_Decembre"=as.Date('12-01-2011', format="%m-%d-%Y"),
                "April.2016"=as.Date('04-01-2016', format="%m-%d-%Y"), "May.2016"=as.Date('05-01-2016', format="%m-%d-%Y"), # Masoala ownership
                "September.2016"=as.Date('09-01-2016', format="%m-%d-%Y"), "January.2017"=as.Date('01-01-2017', format="%m-%d-%Y"),
                "Mar.16"=as.Date('03-01-2016', format="%m-%d-%Y"), "May.16"=as.Date('05-01-2016', format="%m-%d-%Y"), # Vaccination data
                "Sep.16"=as.Date('09-01-2016', format="%m-%d-%Y"), "Jan.17"=as.Date('01-01-2017', format="%m-%d-%Y"),
                "May.17"=as.Date('05-01-2017', format="%m-%d-%Y"), "Sep.17"=as.Date('09-01-2017', format="%m-%d-%Y"),
                "Jan.18"=as.Date('01-01-2018', format="%m-%d-%Y"), "June.18"=as.Date('06-01-2018', format="%m-%d-%Y"),
                "Dec.18"=as.Date('12-01-2018', format="%m-%d-%Y"), "April.19"=as.Date('04-01-2019', format="%m-%d-%Y"),
                "Aug.19"=as.Date('08-01-2019', format="%m-%d-%Y"), "Dec.2019"=as.Date('12-01-2019', format="%m-%d-%Y"),
                "06 zav 2022"=as.Date('01-06-2022', format="%m-%d-%Y"), "07 zav 2022"=as.Date('01-07-2022', format="%m-%d-%Y"),
                "08 zav 2022"=as.Date('01-08-2022', format="%m-%d-%Y"), "09 zav 2022"=as.Date('01-09-2022', format="%m-%d-%Y"),
                "10 zav 2022"=as.Date('01-10-2022', format="%m-%d-%Y"), "11 zav 2022"=as.Date('01-11-2022', format="%m-%d-%Y"),
                "06 zav 2022"=as.Date('01-06-2022', format="%m-%d-%Y"), "16 Avril 2022"=as.Date('04-16-2022', format="%m-%d-%Y"),
                "17 Avril 2022"=as.Date('04-17-2022', format="%m-%d-%Y"), "18 Avril 2022"=as.Date('04-18-2022', format="%m-%d-%Y"),
                "19 Avril 2022"=as.Date('04-19-2022', format="%m-%d-%Y"), "20 Avril 2022"=as.Date('04-20-2022', format="%m-%d-%Y"),
                "21 Avril 2022"=as.Date('04-21-2022', format="%m-%d-%Y"), "22 Avril 2022"=as.Date('04-22-2022', format="%m-%d-%Y"),
                "23 Avril 2022"=as.Date('04-23-2022', format="%m-%d-%Y"), "24 Avril 2022"=as.Date('04-24-2022', format="%m-%d-%Y"),
                "25 Avril 2022"=as.Date('04-25-2022', format="%m-%d-%Y"), "26 Avril 2022"=as.Date('04-26-2022', format="%m-%d-%Y"),
                "27 Avril 2022"=as.Date('04-27-2022', format="%m-%d-%Y"), "28 Avril 2022"=as.Date('04-28-2022', format="%m-%d-%Y"),
                "29 Avril 2022"=as.Date('04-29-2022', format="%m-%d-%Y"), "30 Avril 2022"=as.Date('04-30-2022', format="%m-%d-%Y"),
                "03 Mai 2022"=as.Date('05-03-2022', format="%m-%d-%Y"), "04 Mai 2022"=as.Date('05-04-2022', format="%m-%d-%Y"), 
                "05 Mai 2022"=as.Date('05-05-2022', format="%m-%d-%Y"), "06 Mai 2022"=as.Date('05-06-2022', format="%m-%d-%Y"), 
                "07 Mai 2022"=as.Date('05-07-2022', format="%m-%d-%Y"), "08 Mai 2022"=as.Date('05-08-2022', format="%m-%d-%Y"),
                ch))
}
numeric_clean <- function(dataset) {
  dataset <- as.data.frame(dataset)
  cleaned_dataset_ls <- list()
  to_fix <- c()
  for (i in 1:ncol(dataset)) {
    subbed_col <- gsub('O', '0', 
                       gsub('o', '0', 
                            gsub('No', '0', 
                                 gsub('Tsia', '0', 
                                      gsub('\\+', '', 
                                           gsub('Eny', '1',
                                                gsub('eny', '1', 
                                                     gsub('999', '1', 
                                                          gsub('NO', '0',
                                                               gsub('TSY TETO ANTOREANA', NA, 
                                                                    gsub('M', NA, 
                                                                         gsub('1N', NA, 
                                                                              gsub('1No', '0', 
                                                                                   gsub('tsia', '0', 
                                                                                        gsub('tsiaa', '0', 
                                                                                             gsub('Envy', '1', 
                                                                                                  gsub('Yes', '1', 
                                                                                                       gsub(',No', '0',
                                                                                                            gsub('Tisa', '0',  
                                                                                                                 gsub('ENY', '1', 
                                                                                                                      gsub('tsa', '0', 
                                                                                                                           gsub('Tsy', '0', 
                                                                                                                                gsub('Anjiabe ankofabe', NA,
                                                                                                                                     gsub("Tsy  mbola natao fanadiadiana", NA,
                                                                                                                                          gsub("Eny  tsy tratra antrano", '1', 
                                                                                                                                               gsub("Eny  tsy tratra antrano", '1', 
                                                                                                                                                    gsub('Tsia vaovao', '0', 
                                                                                                                                                         gsub('Sis', NA,
                                                                                                                                                              gsub('mijono', NA, 
                                                                                                                                                                   gsub('2 avy maroantsetra 2avy vinanibe', NA, 
                                                                                                                                                                        gsub('Vinanibe', NA, 
                                                                                                                                                                             gsub('Kopinda', NA, 
                                                                                                                                                                                  gsub('Tsy fantatra', NA,
                                                                                                                                                                                       gsub("Tsy Fantatra", NA,
                                                                                                                                                                                            gsub('Tsy tratra ant rag on', NA,
                                                                                                                                                                                                gsub('50plus', '50', 
                                                                                                                                                                                                    gsub('Mihoatra50', '50', dataset[,i])))))))))))))))))))))))))))))))))))))
    trimmed_subbed_col <- trimws(subbed_col)
    trimmed_subbed_col[which(trimmed_subbed_col == '')] <- NA
    to_fix <- c(to_fix, unique(dataset[which(is.na(as.numeric(trimmed_subbed_col)) != is.na(trimmed_subbed_col)),i]))
    cleaned_dataset_ls[[i]] <-data.frame(matrix(as.numeric(trimmed_subbed_col), ncol=1, nrow=length(trimmed_subbed_col)))
  }
  if (length(unique(to_fix)) != 0) {
    print(unique(to_fix))
    stop('Clean above values.')
  }
  cleaned <- do.call(cbind, cleaned_dataset_ls)
  if (nrow(cleaned) != nrow(dataset) | ncol(cleaned) != ncol(dataset)) {
    stop('Dimensions not correct.')
  }
  return(cleaned)
}
